backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy>=1.26.0
google-cloud-vision==3.4.5
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
pydantic==2.5.0
aiofiles==23.2.1
openai==1.12.0